app/requirements.txt
Django==1.5.1
Fabric==1.8.1
djangorestframework==2.3.12
